making timing more accurate, and adding plans
Will need to implement some cleaning up of old files before new session.
Will also need to finish implementing the mechanism to wait until all
the workers have joined, before we start delegating the work. The idea
is that we won't send back the worker ID to any worker, until we have
all the required number of nodes. (This will probably require starting
the manager and the workers using the same script, specifying the number
of workers as the argument to both programs).
ALSO! The current repository I am cloning is way too big to be
practical. I should instead find a smaller one that doesn't have any
syntax errors in the commit history!