Making the system calculation more robust and accurate
Timing now starts once all the workers are ready to go, and not before
that.
The workers also wait to register until the manager has finished
starting up.

Current issue: trying to delete the existing cloned versions of the
repository upon startup. This is important because it means that the
workers won't be out of sync in terms of startup time - if a worker
already had the repository cloned, it can start up instantly. If an
extra worker is created, it can sometimes still be cloning when the
other worker is finished! The only way to level the playing field is by
deleting all the existing copies on startup: but until then, we have a
problem.